Extract PLC directory functions into plc.ts
Move all functions that call createPlcClient into a new plc.ts file:
- createPlcClient() and PLC_DIRECTORY_URL constant
- submitDID(), updateDID(), and all DID update functions
- getFairAlias() and related error classes (NoAliasError, MultipleAliasesError)

The operation builder functions (updateServiceUrlInOp, addVerificationKeyToOp, etc.)
remain in did.ts as they are pure functions that don't interact with the network.

All moved functions are re-exported from did.ts and domain.ts for backward
compatibility, ensuring no breaking changes for existing code.
